requirements.txt
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
diskcache==5.6.3
Flask==3.1.0
gTTS==2.5.4
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
networkx==3.4.2
numba==0.60.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pydub==0.25.1
PyMatting==1.1.13
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
SpeechRecognition==3.11.0
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
